+++ src/main/java/laddergame/domain/ladder/Line.java
@@ -1,5 +1,6 @@
-package laddergame.domain.rung;
+package laddergame.domain.ladder;
 
+import laddergame.domain.rung.Rung;
 import laddergame.util.BooleanGenerator;
 
 import java.util.ArrayList;
